카카오페이 결제 백엔드 마무리 작업 완료하였고 결제 이후 데이터베이스 결제 정보, 코스 등록까지 되도록 구현했음. 테스트 회…

+++ src/main/java/com/example/ei_backend/repository/LectureRepository.java
@@ -11,6 +11,6 @@ public interface LectureRepository extends JpaRepository<Lecture, Long> {
 
     List<Lecture> findByCourseIdOrderByOrderIndexAsc(Long courseId);
 
-    long countByCourseId(Long courseId);
+    int countByCourseId(Long courseId);
 
 }
